requirements.txt
altair==5.2.0
anyascii==0.3.2
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.7.0
blis==0.7.9
Bottleneck==1.3.5
Brotli==1.0.9
cachetools==5.3.2
catalogue==2.0.7
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
colorclass==2.2.2
comm==0.1.2
confection==0.0.4
contourpy==1.2.0
contractions==0.1.73
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
emoji==2.8.0
en-core-web-sm==3.5.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==0.8.3
Faker==20.1.0
fastjsonschema==2.16.2
favicon==0.7.0
fonttools==4.25.0
gitdb==4.0.11
GitPython==3.1.40
htbuilder==0.6.2
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.0.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.15.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter-server==1.23.4
jupyterlab==3.3.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==10.1.0
munkres==1.1.4
murmurhash==1.0.7
nbclassic==0.5.5
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
neattext==0.1.3
nest-asyncio==1.5.6
nltk==3.8.1
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.0
packaging==23.1
palettable==3.3.3
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
pip-upgrader==1.4.15
platformdirs==3.10.0
plotly==5.9.0
preprocess-kgptalkie==0.1.3
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.25.1
psutil==5.9.0
pure-eval==0.2.2
pyahocorasick==2.0.0
pyarrow==14.0.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==10.5
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==68.0.0
shellingham==1.5.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.1
sniffio==1.2.0
soupsieve==2.5
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
squarify==0.4.3
srsly==2.4.8
st-annotated-text==4.0.1
stack-data==0.2.0
streamlit==1.29.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
terminado==0.17.1
terminaltables==3.1.10
textblob==0.17.1
textsearch==0.0.24
thinc==8.1.10
threadpoolctl==2.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.7.1
typer==0.4.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
vaderSentiment==3.3.2
validators==0.22.0
wasabi==0.9.1
watchdog==3.0.0
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.41.2
win-inet-pton==1.1.0
wordcloud==1.9.2
wordninja==2.0.0
xgboost==2.0.2
zipp==3.11.0